pyproject.toml
[tool.poetry]
package-mode = false
name = "dspy-multi-agent"
version = "0.1.0"
description = ""
authors = ["NixLoar <nix.loar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13.0,<3.14"
python-dotenv = "^1.2.1"
dspy = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = [".venv"]
typeCheckingMode = "off"
venv = ".venv"
venvPath = "."